pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "the-ceo-breach"
version = "0.1.0"
description = "Multi-LLM social engineering simulation"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "openai>=1.0",
    "requests>=2.31",
    "sentence-transformers>=3.0",
    "pandas>=2.0",
    "numpy>=1.26",
    "scipy>=1.12",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "plotly>=5.18",
    "streamlit>=1.30",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
src = ["prompts/**/*.md"]

[tool.pytest.ini_options]
testpaths = ["tests"]
